Fix up naming and minor fixes
