+++ api/channel.go
@@ -72,6 +72,9 @@ type CreateChannelData struct {
 	// ChannelTypes: Voice
 	VoiceQualityMode discord.VideoQualityMode `json:"voice_quality_mode,omitempty"`
 
+	AvailableTags        []discord.Tag          `json:"available_tags,omitempty"`
+	DefaultReactionEmoji *discord.ForumReaction `json:"default_reaction_emoji,omitempty"`
+
 	AuditLogReason `json:"-"`
 }
 
@@ -209,10 +212,11 @@ type ModifyChannelData struct {
 	// non-moderators to a thread; only available on private threads
 	Invitable option.Bool `json:"invitable,omitempty"`
 
-	AuditLogReason `json:"-"`
+	AvailableTags        *[]discord.Tag          `json:"available_tags,omitempty"`
+	AppliedTags          *[]discord.TagID        `json:"applied_tags,omitempty"`
+	DefaultReactionEmoji **discord.ForumReaction `json:"default_reaction_emoji,omitempty"`
 
-	AvailableTags *[]discord.Tag   `json:"available_tags,omitempty"`
-	AppliedTags   *[]discord.TagID `json:"applied_tags,omitempty"`
+	AuditLogReason `json:"-"`
 }
 
 // ModifyChannel updates a channel's settings.
