+++ discord/snowflake_types.go
@@ -340,6 +340,30 @@ func (s StickerPackID) Worker() uint8     { return Snowflake(s).Worker() }
 func (s StickerPackID) PID() uint8        { return Snowflake(s).PID() }
 func (s StickerPackID) Increment() uint16 { return Snowflake(s).Increment() }
 
+// TagID is the snowflake type for a TagID.
+type TagID Snowflake
+
+// NullTagID gets encoded into a null. This is used for optional and nullable snowflake fields.
+const NullTagID = TagID(NullSnowflake)
+
+func (s TagID) MarshalJSON() ([]byte, error)  { return Snowflake(s).MarshalJSON() }
+func (s *TagID) UnmarshalJSON(v []byte) error { return (*Snowflake)(s).UnmarshalJSON(v) }
+
+// String returns the ID, or nothing if the snowflake isn't valid.
+func (s TagID) String() string { return Snowflake(s).String() }
+
+// IsValid returns whether or not the snowflake is valid.
+func (s TagID) IsValid() bool { return Snowflake(s).IsValid() }
+
+// IsNull returns whether or not the snowflake is null. This method is rarely
+// ever useful; most people should use IsValid instead.
+func (s TagID) IsNull() bool { return Snowflake(s).IsNull() }
+
+func (s TagID) Time() time.Time   { return Snowflake(s).Time() }
+func (s TagID) Worker() uint8     { return Snowflake(s).Worker() }
+func (s TagID) PID() uint8        { return Snowflake(s).PID() }
+func (s TagID) Increment() uint16 { return Snowflake(s).Increment() }
+
 // TeamID is the snowflake type for a TeamID.
 type TeamID Snowflake
